Add escapes for separators
This allows you to pass, for example, tabs or NULs as field separators
(these are usually cases where you aren't extracting from mulitple
lines). There are a handful of escapes, but it's basically just the
C ones, but with NUL as \z instead of \0 (to avoid adding an exception
to parsing integers).